+++ test/cli/node/errors_test.dart
@@ -18,7 +18,7 @@ void main() {
   sharedTests(runSass);
 
   test("for package urls", () async {
-    await d.file("test.scss", "@import 'package:nope/test';").create();
+    await d.file("test.scss", "@use 'package:nope/test';").create();
 
     var sass = await runSass(["--no-unicode", "test.scss"]);
     expect(
@@ -26,10 +26,10 @@ void main() {
         emitsInOrder([
           "Error: \"package:\" URLs aren't supported on this platform.",
           "  ,",
-          "1 | @import 'package:nope/test';",
-          "  |         ^^^^^^^^^^^^^^^^^^^",
+          "1 | @use 'package:nope/test';",
+          "  | ^^^^^^^^^^^^^^^^^^^^^^^^",
           "  '",
-          "  test.scss 1:9  root stylesheet"
+          "  test.scss 1:1  root stylesheet"
         ]));
     await sass.shouldExit(65);
   });