+++ test/embedded/file_importer_test.dart
@@ -25,7 +25,7 @@ void main() {
     late OutboundMessage_FileImportRequest request;
 
     setUp(() async {
-      process.send(compileString("@import 'other'", importers: [
+      process.send(compileString("@use 'other'", importers: [
         InboundMessage_CompileRequest_Importer()..fileImporterId = 1
       ]));
 
@@ -64,7 +64,7 @@ void main() {
     late OutboundMessage_FileImportRequest request;
 
     setUp(() async {
-      process.send(compileString("@import 'other'", importers: [
+      process.send(compileString("@use 'other'", importers: [
         InboundMessage_CompileRequest_Importer()..fileImporterId = 1
       ]));
 
@@ -78,7 +78,7 @@ void main() {
             ..id = request.id
             ..fileUrl = ""));
 
-        await _expectImportError(
+        await _expectUseError(
             process, 'The file importer must return an absolute URL, was ""');
         await process.close();
       });
@@ -89,7 +89,7 @@ void main() {
             ..id = request.id
             ..fileUrl = "foo"));
 
-        await _expectImportError(process,
+        await _expectUseError(process,
             'The file importer must return an absolute URL, was "foo"');
         await process.close();
       });
@@ -100,7 +100,7 @@ void main() {
             ..id = request.id
             ..fileUrl = "other:foo"));
 
-        await _expectImportError(process,
+        await _expectUseError(process,
             'The file importer must return a file: URL, was "other:foo"');
         await process.close();
       });
@@ -112,8 +112,7 @@ void main() {
     var importerId = 5679;
     late OutboundMessage_FileImportRequest request;
     setUp(() async {
-      process.send(
-          compileString("@import 'other'", id: compilationId, importers: [
+      process.send(compileString("@use 'other'", id: compilationId, importers: [
         InboundMessage_CompileRequest_Importer()..fileImporterId = importerId
       ]));
       request = await getFileImportRequest(process);
@@ -130,13 +129,13 @@ void main() {
     });
 
     test("whether the import came from an @import", () async {
-      expect(request.fromImport, isTrue);
+      expect(request.fromImport, isFalse);
       await process.kill();
     });
   });
 
   test("errors cause compilation to fail", () async {
-    process.send(compileString("@import 'other'", importers: [
+    process.send(compileString("@use 'other'", importers: [
       InboundMessage_CompileRequest_Importer()..fileImporterId = 1
     ]));
 
@@ -148,13 +147,13 @@ void main() {
 
     var failure = await getCompileFailure(process);
     expect(failure.message, equals('oh no'));
-    expect(failure.span.text, equals("'other'"));
-    expect(failure.stackTrace, equals('- 1:9  root stylesheet\n'));
+    expect(failure.span.text, equals("@use 'other'"));
+    expect(failure.stackTrace, equals('- 1:1  root stylesheet\n'));
     await process.close();
   });
 
   test("null results count as not found", () async {
-    process.send(compileString("@import 'other'", importers: [
+    process.send(compileString("@use 'other'", importers: [
       InboundMessage_CompileRequest_Importer()..fileImporterId = 1
     ]));
 
@@ -165,13 +164,13 @@ void main() {
 
     var failure = await getCompileFailure(process);
     expect(failure.message, equals("Can't find stylesheet to import."));
-    expect(failure.span.text, equals("'other'"));
+    expect(failure.span.text, equals("@use 'other'"));
     await process.close();
   });
 
   group("attempts importers in order", () {
     test("with multiple file importers", () async {
-      process.send(compileString("@import 'other'", importers: [
+      process.send(compileString("@use 'other'", importers: [
         for (var i = 0; i < 10; i++)
           InboundMessage_CompileRequest_Importer()..fileImporterId = i
       ]));
@@ -188,7 +187,7 @@ void main() {
     });
 
     test("with a mixture of file and normal importers", () async {
-      process.send(compileString("@import 'other'", importers: [
+      process.send(compileString("@use 'other'", importers: [
         for (var i = 0; i < 10; i++)
           if (i % 2 == 0)
             InboundMessage_CompileRequest_Importer()..fileImporterId = i
@@ -218,9 +217,9 @@ void main() {
 
   test("tries resolved URL as a relative path first", () async {
     await d.file("upstream.scss", "a {b: c}").create();
-    await d.file("midstream.scss", "@import 'upstream';").create();
+    await d.file("midstream.scss", "@use 'upstream';").create();
 
-    process.send(compileString("@import 'midstream'", importers: [
+    process.send(compileString("@use 'midstream'", importers: [
       for (var i = 0; i < 10; i++)
         InboundMessage_CompileRequest_Importer()..fileImporterId = i
     ]));
@@ -251,7 +250,7 @@ void main() {
     });
 
     test("without a base URL", () async {
-      process.send(compileString("@import 'other'",
+      process.send(compileString("@use 'other'",
           importer: InboundMessage_CompileRequest_Importer()
             ..fileImporterId = 1));
 
@@ -268,7 +267,7 @@ void main() {
     });
 
     test("with a base URL", () async {
-      process.send(compileString("@import 'other'",
+      process.send(compileString("@use 'other'",
           url: p.toUri(d.path("input")).toString(),
           importer: InboundMessage_CompileRequest_Importer()
             ..fileImporterId = 1));
@@ -281,8 +280,8 @@ void main() {
 
 /// Asserts that [process] emits a [CompileFailure] result with the given
 /// [message] on its protobuf stream and causes the compilation to fail.
-Future<void> _expectImportError(EmbeddedProcess process, Object message) async {
+Future<void> _expectUseError(EmbeddedProcess process, Object message) async {
   var failure = await getCompileFailure(process);
   expect(failure.message, equals(message));
-  expect(failure.span.text, equals("'other'"));
+  expect(failure.span.text, equals("@use 'other'"));
 }