+++ test/embedded/protocol_test.dart
@@ -68,7 +68,7 @@ void main() {
     });
 
     test("caused by duplicate compilation IDs", () async {
-      process.send(compileString("@import 'other'", importers: [
+      process.send(compileString("@use 'other'", importers: [
         InboundMessage_CompileRequest_Importer()..importerId = 1
       ]));
       await getCanonicalizeRequest(process);