+++ pkg/sass_api/CHANGELOG.md
@@ -1,4 +1,8 @@
-## 13.0.2
+## 13.1.1
+
+* No user-visible changes.
+
+## 13.1.0
 
 * No user-visible changes.